+++ package.json
@@ -5,8 +5,9 @@
   "main": "index.js",
   "scripts": {
     "test": "jest",
-    "lint": "tslint include/ts/*.ts --exclude include/ts/*.d.ts",
-    "prettier": "prettier --write 'include/**/*.ts'",
+    "test:watch":"jest --watch",
+    "lint": "tslint src/*.ts --exclude src/*.d.ts",
+    "prettier": "prettier --write 'src/**/*.ts'",
     "build": "tsc",
     "tsc:watch": "tsc -w"
   },
@@ -14,36 +15,19 @@
     "type": "git",
     "url": "git+https://github.com/danieljharvey/it-is-the-egg.git"
   },
-  "keywords": [
-    "Egg"
-  ],
+  "keywords": ["Egg"],
   "author": "Daniel J Harvey",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/danieljharvey/it-is-the-egg/issues"
   },
   "homepage": "https://github.com/danieljharvey/it-is-the-egg#readme",
-  "devDependencies": {
-    "@types/core-js": "^0.9.43",
-    "@types/jest": "^20.0.5",
-    "@types/ramda": "^0.24.7",
-    "jest": "^20.0.4",
-    "prettier": "^1.6.1",
-    "ts-jest": "^20.0.7",
-    "tslint": "^5.8.0",
-    "tslint-config-prettier": "^1.5.0",
-    "typescript": "^2.4.2"
-  },
   "jest": {
     "transform": {
       ".(ts|tsx)": "<rootDir>/node_modules/ts-jest/preprocessor.js"
     },
     "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js"
-    ]
+    "moduleFileExtensions": ["ts", "tsx", "js"]
   },
   "dependencies": {
     "@types/es6-promise": "^0.0.33",
@@ -58,6 +42,15 @@
     "lodash": "^4.17.4",
     "ramda": "^0.24.1",
     "ramda-ts": "^0.1.2",
-    "tsmonad": "^0.8.0"
+    "tsmonad": "^0.8.0",
+    "@types/core-js": "^0.9.43",
+    "@types/jest": "^20.0.5",
+    "@types/ramda": "^0.24.7",
+    "jest": "^20.0.4",
+    "prettier": "^1.6.1",
+    "ts-jest": "^20.0.7",
+    "tslint": "^5.8.0",
+    "tslint-config-prettier": "^1.5.0",
+    "typescript": "^2.4.2"
   }
 }
