+++ src/dom/__tests__/CanvasTest.ts
