+++ src/logic/TheEgg.ts
