+++ src/logic/__tests__/PathFinderTest.ts
