+++ src/objects/__tests__/PlayerTest.ts
